+++ test/common/http/header_utility_test.cc
@@ -13,8 +13,7 @@ void envoy_test_release(void* context) {
 }
 
 envoy_data envoyTestString(std::string& s, uint32_t* sentinel) {
-  return {static_cast<int>(s.size()), reinterpret_cast<const uint8_t*>(s.c_str()),
-          envoy_test_release, sentinel};
+  return {s.size(), reinterpret_cast<const uint8_t*>(s.c_str()), envoy_test_release, sentinel};
 }
 
 TEST(HeaderDataConstructorTest, FromCToCppEmpty) {
